.develop/spec.json,.develop/test.sh,go.mod,rpc: init working impl
